requirements.txt
Django~=4.2.3
djangorestframework~=3.14.0
django-filter~=23.2
djangochannelsrestframework~=1.1.0
celery~=5.3.1
requests~=2.31.0
channels~=4.0.0